+++ src/lib.rs
@@ -61,7 +61,7 @@ mod routing_table;
 mod routing_node;
 mod routing_core;
 mod relay;
-mod who_are_you;
+mod hello;
 
 pub mod routing;
 pub mod refresh_accumulator;
